refactor: LottoPrizeNumbers 클래스 생성
* LottoTicketMachine의 역할이 너무 무거워 당첨번호, 보너스번호를 저장한 클래스 생성 후 역할 분리
