+++ src/main/java/lotto/Application.java
@@ -1,8 +1,12 @@
 package lotto;
 
+import java.util.List;
+
 import lotto.domains.customer.Customer;
+import lotto.domains.lotto.LottoPrizeNumbers;
 import lotto.domains.lotto.LottoTicket;
 import lotto.domains.lotto.LottoTicketMachine;
+import lotto.service.LottoService;
 import lotto.util.TypeConverter;
 import lotto.view.InputInterface;
 import lotto.view.OutputInterface;
@@ -10,8 +14,9 @@ import lotto.view.OutputInterface;
 public class Application {
 	public static void main(String[] args) {
 		InputInterface inputInterface = new InputInterface();
-		Customer customer = purchaseLottoTickets(inputInterface);
+		LottoService lottoService = new LottoService();
 
+		Customer customer = purchaseLottoTickets(inputInterface);
 		int ticketAmount = customer.calculateAmount();
 		OutputInterface.printMessage(ticketAmount + OutputInterface.PURCHASE_AMOUNT.toString());
 
@@ -21,20 +26,24 @@ public class Application {
 		OutputInterface.printMessage(tickets.toString());
 		OutputInterface.printNewLine();
 
-		drawWinningNumbers(inputInterface, lottoTicketMachine);
+		List<Integer> winningNumbers = drawWinningNumbers(inputInterface, lottoService);
+		int bonusNumber = drawBonusNumber(inputInterface, lottoService);
 
+		LottoPrizeNumbers lottoPrizeNumbers = LottoPrizeNumbers.of(winningNumbers, bonusNumber);
 
 	}
 
-	private static void drawWinningNumbers(InputInterface inputInterface,
-		LottoTicketMachine lottoTicketMachine) {
+	private static int drawBonusNumber(InputInterface inputInterface, LottoService lottoService) {
+
+	}
+
+	private static List<Integer> drawWinningNumbers(InputInterface inputInterface, LottoService lottoService) {
 		while (true) {
 			try {
 				OutputInterface.printMessage(OutputInterface.ENTER_WINNING_NUMBERS);
 				String winningNumbers = inputInterface.readLine();
 
-				lottoTicketMachine.drawWinningNumbers(winningNumbers);
-				break;
+				return lottoService.drawWinningNumbers(winningNumbers);
 			} catch (IllegalArgumentException exception) {
 				processException(exception);
 			}
