+++ src/main/java/lotto/domains/lotto/LottoPrizeNumbers.java
@@ -1,0 +1,17 @@
+package lotto.domains.lotto;
+
+import java.util.List;
+
+public class LottoPrizeNumbers {
+	private List<Integer> winningNumbers;
+	private int bonusNumber;
+
+	private LottoPrizeNumbers(List<Integer> winningNumbers, int bonusNumber) {
+		this.winningNumbers = winningNumbers;
+		this.bonusNumber = bonusNumber;
+	}
+
+	public static LottoPrizeNumbers of(List<Integer> winningNumbers, int bonusNumber) {
+		return new LottoPrizeNumbers(winningNumbers, bonusNumber);
+	}
+}
